fix avx2 path error
